feat: add endpoints to allow subscriptions management by user

+++ src/helpers/utils.ts
@@ -1,4 +1,5 @@
 import db from './mysql';
+import { SUBSCRIPTIONS_TYPE } from '../templates';
 import type { Response } from 'express';
 
 export function rpcSuccess(res: Response, result: string, id: string | number) {
@@ -21,6 +22,12 @@ export function rpcError(res: Response, code: number, e: unknown, id: string | n
   });
 }
 
+export function sanitizeSubscriptions(list?: any) {
+  return (Array.isArray(list) ? list : [list]).filter((item: any) =>
+    SUBSCRIPTIONS_TYPE.includes(item)
+  ) as typeof SUBSCRIPTIONS_TYPE;
+}
+
 export async function subscribe(email: string, address: string) {
   const created = (Date.now() / 1e3).toFixed();
   const subscriber = { email, address, created };
@@ -35,8 +42,32 @@ export async function verify(email: string, address: string) {
   );
 }
 
-export async function unsubscribe(email: string) {
-  return await db.queryAsync('DELETE FROM subscribers WHERE email = ?', [email]);
+export async function update(email: string, address: string, subscriptions: string[]) {
+  const subs = sanitizeSubscriptions(subscriptions);
+  if (subs.length === 0) {
+    return await db.queryAsync('DELETE FROM subscribers WHERE email = ? and address = ?', [
+      email,
+      address
+    ]);
+  } else {
+    return await db.queryAsync(
+      'UPDATE subscribers SET subscriptions = ? WHERE email = ? AND address = ? LIMIT 1',
+      [JSON.stringify(sanitizeSubscriptions(subscriptions)), email, address]
+    );
+  }
+}
+
+export async function unsubscribe(email: string, subscriptions?: string[]) {
+  const subs = sanitizeSubscriptions(subscriptions);
+
+  if (subs.length === 0) {
+    return await db.queryAsync('DELETE FROM subscribers WHERE email = ?', [email]);
+  } else {
+    return await db.queryAsync('UPDATE subscribers SET subscriptions = ? WHERE email = ?', [
+      JSON.stringify(subs),
+      email
+    ]);
+  }
 }
 
 export async function getEmailAddresses(email: string) {
@@ -45,8 +76,18 @@ export async function getEmailAddresses(email: string) {
   ]);
 }
 
-export async function getUniqueEmails() {
-  return await db.queryAsync('SELECT DISTINCT email FROM subscribers WHERE verified > 0');
+export async function getUniqueEmails(subscriptionType: string) {
+  return await db.queryAsync(
+    `SELECT DISTINCT email FROM subscribers WHERE verified > 0 AND JSON_CONTAINS(subscriptions, '${JSON.stringify(
+      sanitizeSubscriptions(subscriptionType)[0]
+    )}')`
+  );
+}
+
+export async function getAddressSubscriptions(address: string) {
+  return await db.queryAsync('SELECT subscriptions from subscribers WHERE address = ? LIMIT 1', [
+    address
+  ]);
 }
 
 // RFC5322 standard, does support most format, but not all
